+++ data_hub_api/main.py
@@ -1,5 +1,4 @@
 import logging
-import uuid
 
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
@@ -11,6 +10,7 @@ from data_hub_api.kotahi_docmaps.v1.api_router import (
 from data_hub_api.utils.cache import InMemorySingleObjectCache
 from data_hub_api.docmaps.v2.provider import DocmapsProvider
 from data_hub_api.kotahi_docmaps.v1.provider import DocmapsProvider as KotahiDocmapsProvider
+from data_hub_api.utils.fastapi import log_requests_middleware
 
 
 LOGGER = logging.getLogger(__name__)
@@ -19,28 +19,7 @@ LOGGER = logging.getLogger(__name__)
 def create_app():
     app = FastAPI()
 
-    @app.middleware("http")
-    async def log_requests(request, call_next):
-        request_id = str(uuid.uuid4())
-        user_agent = request.headers.get("user-agent", "unknown")
-        client_ip = request.client.host if request.client else "unknown"
-        protocol = request.scope.get("http_version", "unknown")
-        method = request.method
-        path = request.url.path
-
-        request.state.request_id = request_id
-
-        LOGGER.info(
-            f'{request_id} START {client_ip} - - "{method} {path} HTTP/{protocol}" "{user_agent}"'
-        )
-
-        response = await call_next(request)
-
-        LOGGER.info(
-            f'{request_id} END {client_ip} - - "{method} {path} HTTP/{protocol}" {response.status_code} "{user_agent}"'
-        )
-
-        return response
+    app.middleware("http")(log_requests_middleware())
 
     docmaps_v2_max_age_in_seconds = 10 * 60  # 10 minutes
     kotahi_docmaps_max_age_in_seconds = 60 * 60  # 1 hour
